Finishing the login for returning members -Members are selected from the list -Members are confirmed, declined, or updated -Members generate visit records
Features to be added later
-Search feature that sorts members based on the characters added first.

+++ app/src/main/java/generalinfo/powellmakerspace/org/makerspacelogin/Member.java
@@ -25,7 +25,7 @@ public class Member {
     }
 
     // Setters
-    public void setMemberID(long memberID){
+    public void setMemberID(int memberID){
         this.memberID = memberID;
     }
